fix: Resolve the multipart-form decoding issue
It turns out that the `req.body` will not be available unless it has
been parsed.In our case, since the first object was the file, the rest
of elements were not parsed and hence weren't available for use.  Care
should be taken to make sure that the file to be uploaded is the last
item in the form-data list.